+++ src/main/java/com/bitbank/repository/v1/BankAccountTransactionsRepository.java
@@ -1,5 +1,8 @@
 package com.bitbank.repository.v1;
 
+import java.math.BigDecimal;
+
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -7,5 +10,8 @@ import com.bitbank.entities.v1.BankAccountTransactionsEntity;
 
 @Repository
 public interface BankAccountTransactionsRepository extends CrudRepository<BankAccountTransactionsEntity, Long> {
+
+	@Query("SELECT SUM(b.amount) FROM BankAccountTransactionsEntity b")
+	BigDecimal balance();
 
 }