+++ src/main/java/com/bitbank/services/v1/BankAccountTransactionsService.java
@@ -1,5 +1,6 @@
 package com.bitbank.services.v1;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -11,29 +12,39 @@ import com.bitbank.repository.v1.BankAccountTransactionsRepository;
 
 @Service
 public class BankAccountTransactionsService {
-	
+
 	@Autowired
 	private BankAccountTransactionsRepository bankAccountTransactionsRepository;
-	
+
 	/**
 	 * Index
+	 * 
 	 * @return
 	 */
 	public List<BankAccountTransactionsEntity> index() {
 		Iterable<BankAccountTransactionsEntity> items = bankAccountTransactionsRepository.findAll();
 		ArrayList<BankAccountTransactionsEntity> bankAccountTransactions = new ArrayList<>();
-		
-		for (BankAccountTransactionsEntity transaction:items) {
+
+		for (BankAccountTransactionsEntity transaction : items) {
 			bankAccountTransactions.add(transaction);
 		}
-		
+
 		return bankAccountTransactions;
 	}
-	
+
 	/**
 	 * Post
 	 */
 	public BankAccountTransactionsEntity post(BankAccountTransactionsEntity bankAccountTransactionsEntity) {
 		return bankAccountTransactionsRepository.save(bankAccountTransactionsEntity);
+	}
+
+	/**
+	 * Return balance
+	 * 
+	 * @return
+	 */
+	public BigDecimal balance() {
+		return bankAccountTransactionsRepository.balance();
 	}
 }